+++ lib/widgets/layer_widget.dart
@@ -314,7 +314,6 @@ class _LayerWidgetState extends State<LayerWidget>
     var layer = _layer as TextLayerData;
     var style = TextStyle(
       fontSize: fontSize * layer.fontScale,
-      fontWeight: FontWeight.w400,
       color: layer.color,
       overflow: TextOverflow.ellipsis,
     );